# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a compile and execute system.",
        "title": "CompileSys",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/result/{task_id}": {
            "get": {
                "description": "\"Get task execute result\"",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task uuid",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.ResultResponse"
                        }
                    },
                    "default": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/status/{task_id}": {
            "get": {
                "description": "\"Get task process GetStatus\"",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get GetStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task uuid",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.StatusResponse"
                        }
                    },
                    "default": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/task": {
            "post": {
                "description": "\"Create task for compile and execute. Need code and compiler name.\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Create task and process",
                "parameters": [
                    {
                        "description": "data for execute",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.PostTaskRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "task uuid",
                        "schema": {
                            "$ref": "#/definitions/types.PostTaskResponse"
                        }
                    },
                    "default": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "types.PostTaskRequestBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "compiler_name": {
                    "type": "string"
                }
            }
        },
        "types.PostTaskResponse": {
            "type": "object",
            "properties": {
                "task_id": {
                    "type": "string"
                }
            }
        },
        "types.ResultResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
        "types.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    }
}